+++ dffml/operation/io.py
@@ -12,7 +12,7 @@ from dffml.df.base import (
 
 # Definitions
 UserInput = Definition(name="UserInput", primitive="str")
-DataToPrint = Definition(name="DataToPrint", primitive="str")
+DataToPrint = Definition(name="DataToPrint", primitive="generic")
 
 AcceptUserInput = Operation(
     name="AcceptUserInput",
@@ -25,23 +25,19 @@ AcceptUserInput = Operation(
 class AcceptUserInputContext(OperationImplementationContext):
     @staticmethod
     def receive_input():
+        print("Enter the value: ", end="")
         return input()
 
     async def run(self, inputs: Dict[str, Any]) -> Dict[str, Any]:
         user_input = await self.parent.loop.run_in_executor(
             self.parent.pool, self.receive_input
         )
-        return {"InputData": {"data": user_input}}
+        return {"InputData": user_input}
 
 
 class AcceptUserInput(OperationImplementation):
     """
     Accept input from stdin using python input()
-
-    Parameters
-    ++++++++++
-    inputs : dict
-        A dictionary with a key and empty list as value.
 
     Returns
     +++++++
@@ -52,21 +48,22 @@ class AcceptUserInput(OperationImplementation):
     ++++++++
 
     The following example shows how to use AcceptUserInput.
+    (Assumes that the input from stdio is "Data flow is awesome"!)
 
     >>> dataflow = DataFlow.auto(AcceptUserInput, GetSingle)
     >>> dataflow.seed.append(
     ...     Input(
     ...         value=[AcceptUserInput.op.outputs["InputData"].name],
-    ...         definition=GetSingle.op.inputs["spec"]
+    ...         definition=GetSingle.op.inputs["spec"],
     ...     )
     ... )
     >>>
     >>> async def main():
-    ...     async for ctx, results in MemoryOrchestrator.run(dataflow, {"input":[]}):
+    ...     async for ctx, results in MemoryOrchestrator.run(dataflow, {"input": []}):
     ...         print(results)
     >>>
     >>> asyncio.run(main())
-    {'UserInput': {'data': 'Data flow is awesome'}}
+    Enter the value: {'UserInput': 'Data flow is awesome'}
     """
 
     op = AcceptUserInput
@@ -92,33 +89,32 @@ class AcceptUserInput(OperationImplementation):
 
 
 @op(inputs={"data": DataToPrint}, outputs={}, conditions=[])
-async def print_output(data: str):
+async def print_output(data: Any):
     """
     Print the output on stdout using python print()
 
     Parameters
     ++++++++++
-    inputs : list
-        A list of Inputs whose value is to be printed.
+    data : Any
+        A python literal to be printed.
 
     Examples
     ++++++++
 
     The following example shows how to use print_output.
 
-    >>> dataflow = DataFlow.auto(print_output, GetSingle)
+    >>> dataflow = DataFlow.auto(print_output)
     >>> inputs = [
     ...     Input(
-    ...         value="print_output example",
-    ...         definition=dataflow.definitions["DataToPrint"],
-    ...         parents=None,)]
+    ...         value="print_output example", definition=print_output.op.inputs["data"]
+    ...     )
+    ... ]
     >>>
     >>> async def main():
     ...     async for ctx, results in MemoryOrchestrator.run(dataflow, inputs):
-    ...         print("String to be printed is 'print_output example'")
+    ...         pass
     >>>
     >>> asyncio.run(main())
     print_output example
-    String to be printed is 'print_output example'
     """
     print(data)