+++ examples/io/dataflow_diagram.sh
@@ -1,0 +1,1 @@
+dffml service dev export examples.io.io_usage:dataflow | dffml dataflow diagram -stages processing -config json /dev/stdin